requirements.txt
astroid==2.6.2
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
configparser==5.0.2
debtcollector==2.3.0
idna==2.10
importlib-metadata==4.8.2
importlib-resources==5.4.0
iso8601==1.0.0
isort==5.9.1
keystoneauth1==4.4.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
msgpack==1.0.2
netaddr==0.8.0
netifaces==0.11.0
os-service-types==1.7.0
oslo.config==8.7.1
oslo.i18n==5.1.0
oslo.serialization==4.2.0
oslo.utils==4.11.0
packaging==21.2
pbr==5.7.0
progressbar2
pylint==2.9.3
pyodk==0.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-keystoneclient==4.3.0
python-swiftclient==3.12.0
pytz==2021.3
PyYAML==6.0
requests==2.25.1
rfc3986==1.5.0
six==1.16.0
stevedore==3.5.0
toml==0.10.2
typed-ast==1.4.3
typer==0.3.2
typing-extensions==3.10.0.0
urllib3==1.26.6
wrapt==1.12.1
zipp==3.6.0
